+++ packages/csc-ui/src/components/c-data-table/c-data-table.tsx
@@ -247,6 +247,7 @@ export class CDataTable {
   async clearSelections() {
     this._selections = [];
     this._emitChange();
+    this._setIndeterminateStatus();
   }
 
   /**