+++ packages/csc-ui/src/components/c-icon/c-icon.tsx
@@ -26,6 +26,13 @@ export class CIcon {
    */
   @Prop() color = 'currentColor';
 
+  private _pathElement: SVGElement;
+
+  componentDidLoad() {
+    // workaround to make the icon work as a c-option child
+    this._pathElement.setAttribute('d', this.host.dataset.path);
+  }
+
   render() {
     return (
       <Host
@@ -33,9 +40,10 @@ export class CIcon {
           '--_c-icon-size': `${this.size}px`,
           '--_c-icon-default-color': this.color,
         }}
+        data-path={this.path}
       >
         <svg viewBox="0 0 24 24">
-          <path d={this.path} />
+          <path ref={(el) => (this._pathElement = el)} />
         </svg>
       </Host>
     );