+++ packages/csc-ui/src/components/c-text-field/c-text-field.tsx
@@ -92,13 +92,6 @@ export class CTextField {
   @Prop() name: string;
 
   /**
-   * Numeric input
-   *
-   * @deprecated Use type="number" instead
-   */
-  @Prop() number = false;
-
-  /**
    * Placeholder of the input
    */
   @Prop() placeholder: string;
@@ -131,7 +124,7 @@ export class CTextField {
   /**
    * Type of the input
    */
-  @Prop() type: string;
+  @Prop() type = 'text';
 
   /**
    * Set the validity of the input
@@ -209,6 +202,10 @@ export class CTextField {
   }
 
   private _handleChange = (event) => {
+    if (this.type === 'number' && !event.target.value && event.key === '-') {
+      return;
+    }
+
     this.value = this.trimWhitespace
       ? event.target.value.trim()
       : event.target.value;
@@ -242,7 +239,7 @@ export class CTextField {
         onBlur: this._handleBlur,
       },
       input: {
-        type: this.number ? 'number' : this.type || 'text',
+        type: this.type,
         min: this.min,
         max: this.max,
         step: this.step,